Enable limit to be 0
It is currently not possible to supply a limit of 0. The following query:

```
squel.select().from('myTable').limit(0)
```

**Before:**
`SELECT * FROM myTable`

**After:**
`SELECT * FROM myTable LIMIT 0`

**Why is this useful?**
Some APIs return responses like the following:
```
{
    data: [{...}, {...}, ...],
    count: 144
}
```
`data` will contain a paginated response (eg. max 20),
`count` will be the total number of results.

If a developer only wants the number of results (count), and not the actual results he will make a request, with a limit of 0, which will return:

```
{
    data: [],
    count: 144
}
```