Add type definitions for Squel. (#322)
* Add type definitions for Squel.

This commit adds complete, documented and and strong type definitions
for Squel.

Closes hiddentao/squel#173, hiddentao/squel#206

* Add type definitions for blocks and small fixes.

This commit completes the type definitions by adding the methods and
properties of the blocks, even if they are not exposed to the query
builders.
This also adds the missing `isSquelBuilder` function and type
parameters for the MySQL flavour extension of Squel.

* Fix typo in type definitions

The property `_onConflict` was typed as `_onCondlict`.

Reported by @alexturek in #322.

* Define parameters for `str` and `rstr` in type definitions

The functions `str` and `rstr` were missing their parameters. This
commit adds these parameters.

Reported by @blocka in #322.